tor-dirserver: Implement request function
This commit adds two functions:
* `request_single`
* `request_mutli`

The first function implements downloading a `tor_dirclient::Requestable`
from a single authority, failing in the case of a failure.

The second function wraps the previous one in a retry loop, utilizing
the `RetryDelay` from `tor-basic-utils` in order to implement it in a
specification compliant fashion.