Move SaturatingSystemTime into tor-basic-utils
This commit moves SaturatingSystemTime into tor-basic-utils and performs
a refactoring on it.  From now on, it no longer uses a hardcoded value
but rather calculates it during run-time inside a LazyLock, allowing for
more overall flexibility.

Also, it implements a new type in tor_dirserver::database, namely
Timestamp, which wraps a SaturatingSystemTime into a new type
implementing FromSql and ToSql, as those trait on the original
SaturatingSystemTime found in tor-dirserver had to be removed, given
that they do not make much sense inside tor-basic-utils, due to SQLite
being a very domain specific thing here.