tor-dirserver: Better tor_dirclient error handling
Previously, we only mapped tor_dirclient::Error::RequestFailed to
tor_dirserver::err::AuthorityCommunicationError::RequestFailed, and all
other entries to the bug counterpart, which is technically not clean, as
it would also bug on future entries.

This commit modifies the error variant to leave the bigger
tor_dirclient::Error as it is.