parse2: ItemValueParseable, netdoc(rest, with): rework
The path in `with` must always be a module, not a function, because we
need to use it for encoding too.

In the case of `rest`, it's probably sensible to require a different
function name, given that the semantics are subtly different to those
of `rest` without `with`, since the latter can only get a single word
but with `rest` you get the whole line.